adds a childaction for the rooms when searching for a room to see the eventoverview of the room
